app/requirements.txt
django-environ<=0.12.0
djangorestframework<=3.16.1
djangorestframework-simplejwt[crypto]<=5.5.1
drf-spectacular[sidecar]==0.28.0
pytest-django<=4.11.1
